Allow larger linearAdvance
Instead of baling when linearAdvance is too big to fit into a short, make Glyphe::linearAdvance an int so it can accept the assignment from info.linearAdvance.

Increasing the size of that member won't affect the sizeof(Glyph):

https://godbolt.org/z/Wo9K8od7n

+++ src/gui/text/qfontengine_p.h
@@ -88,7 +88,7 @@ public:
     struct Glyph {
         Glyph() = default;
         ~Glyph() { delete [] data; }
-        short linearAdvance = 0;
+        int linearAdvance = 0;
         unsigned short width = 0;
         unsigned short height = 0;
         short x = 0;
